requirements.txt
pdfplumber>=0.11.0
Jinja2>=3.1.0
spacy>=3.7.0
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
numpy>=1.26.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
black>=24.0.0
ruff>=0.4.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
streamlit>=1.35.0
streamlit-extras>=0.4.0
requests>=2.31.0
anthropic>=0.25.0
openai>=1.0.0
# Phase 4 — Database + Auth
sqlalchemy[asyncio]>=2.0
alembic>=1.13
asyncpg>=0.29
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt>=4.0,<4.2  # passlib 1.7.4 is incompatible with bcrypt 4.2+
email-validator>=2.0
# Phase 4 — Async testing
pytest-asyncio>=0.23
anyio>=4.0
